Bug 1784880 - Support range requests on blob URLs in fetch/XMLHttpRequest; r=dlrobertson,necko-reviewers,jesup
Differential Revision: https://phabricator.services.mozilla.com/D184281

UltraBlame original commit: 384e4f88f047f5f0797df9577bbbad6ff99ae64c

+++ dom/base/nsContentUtils.cpp
@@ -7521,8 +7521,8 @@ Maybe<nsContentUtils::ParsedRange> nsContentUtils::ParseSingleRangeRequest(
     const nsACString& aHeaderValue, bool aAllowWhitespace) {
   
   mozilla::Tokenizer p(aHeaderValue);
-  Maybe<uint32_t> rangeStart;
-  Maybe<uint32_t> rangeEnd;
+  Maybe<uint64_t> rangeStart;
+  Maybe<uint64_t> rangeEnd;
 
   
   if (!p.CheckWord("bytes")) {
@@ -7545,7 +7545,7 @@ Maybe<nsContentUtils::ParsedRange> nsContentUtils::ParseSingleRangeRequest(
   }
 
   
-  int32_t res;
+  uint64_t res;
   if (p.ReadInteger(&res)) {
     rangeStart = Some(res);
   }
